WATCHMAKER: Move GL Texture logic to its own file
